React Accordion Component update
