+++ react-accordion-component/src/accordion.jsx
@@ -1,43 +1,30 @@
 /* eslint-disable no-console */
 import React from 'react';
-import topics from './topics';
 
-class Collapse extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = { open: false };
-    this.togglePanel = this.togglePanel.bind(this);
-  }
-
-  togglePanel() {
-    this.setState({ open: !this.state.open });
-  }
-
-  render() {
-    return (
-      <div>
-        <div onClick={this.togglePanel} className='header'>
-          {this.props.title}</div>
-        {this.state.open ? (
-          <div className='content'>
-            {this.props.children}
-          </div>
-        ) : null}
-      </div>
-    );
-  }
-}
 
 export default class Accordion extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = { isClicked: null,
+    openId: 0};
+    this.handleClick = this.handleClick.bind(this);
+  }
+
+  handleClick(event) {
+    this.setState({ openId: event });
+  }
+
   render() {
+    const element = this.props.topics.map(topics => {
+      return (
+        <div event={topics.id} >
+          <div event={topics.id} className='header' onClick={() => this.handleClick(topics.id)}>{topics.title}</div>
+          {this.state.openId === topics.id ? (<div className='content'>{topics.content}</div>) : null }
+        </div>
+      );
+    });
     return (
-      <div>
-        <Collapse topics={topics} title={topics.title}>
-          <div>
-            <p>{topics.content}</p>
-          </div>
-        </Collapse>
-      </div>
+      <div>{element}</div>
     );
   }
 }
