--- react-accordion-component/src/topics.jsx
@@ -1,16 +1,0 @@
-const topics = [
-  {
-    title: 'HyperText Markup Lanugage',
-    content: 'Hypertext Markup Language (HTML) is the standard Markup language for creating web pages and web applications.With Cascading Style Sheets(CSS) and Javascript, it forms a triad of cornerstone technologies for the World Wide Web'
-  },
-  {
-    title: 'Cascading Style Sheets',
-    content: 'Cascading Style Sheets (CSS) is a style sheet language used for describing the presentation of a document written in a markup language like HTML. CSS is a cornerstone technology of the World Wide Web alongside HTML and Javascript.'
-  },
-  {
-    title: 'JavaScript',
-    content: 'Javascript, often abbreviated as JS, is a high-level, interpreted programming language that conforms to the ECMAScript specification. JavaScript has curly-bracket syntax, dynamic typing, prototype-based object-orientation, and first-class functions.'
-  }
-];
-
-export default topics;
